RegressionTests

Documentation for RegressionTests.

RegressionTests.differentiate

RegressionTests.runbenchmarks

RegressionTests.test

RegressionTests.differentiate — Method

differentiate(f, g)

Determine if f() and g() have different distributions.

Returns false if f and g have the same distribution and usually returns true if they have different distributions.

Properties

If f and g have the same distribution, the probability of a false positive is 1e-10.

If f and g have distributions that differ* by at least .05, then the probability of a false negative is .05.

If f and g have the same distribution, f and g will each be called 53 times, on average

f and g will each be called at most 300 times.

The difference between distributions is quantified as the integral from 0 to 1 of (cdf(g)(invcdf(f)(x)) - x)^2 where cdf and invcdf are higher order functions that compute the cumulative distribution function and inverse cumulative distribution function, respectively.

More generally, for any k > .025, recall loss is according to empirical estimation, at most max(1e-4, 20^(1-k/.025)). So, for example, a regression with k = .1, will be escape detection at most 1 out of 8000 times.

RegressionTests.runbenchmarks — Method

runbenchmarks()

When called in test/runtests.jl while pwd() points to test, this function runs the benchmarks in bench/runbenchmarks.jl.

There are some keyword arguments, but they are not public.

RegressionTests.test — Method

test(skip_unsupported_platforms=false)

When called in testing, runs regression tests, reports all changes, and throws if there are regressions.

Set skip_unsupported_platforms to true to skip the test (quietly pass) on platforms that are not supported.